update eslint rules to follow wpromote standard
add ability to specify cognitoform version
