+++ .prettierrc
@@ -1,0 +1,9 @@
+{
+  "endOfLine": "auto",
+  "printWidth": 80,
+  "singleQuote": true,
+  "trailingComma": "es5",
+  "semi": true,
+  "jsxSingleQuote": true,
+  "useTabs": true
+}
